requirements.txt
playwright==1.45.0
python-dotenv==1.0.1
PySide6==6.7.1
fastapi==0.114.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.8.2
pytest==8.3.2
